+++ src/distance.rs
@@ -1,15 +1,17 @@
 //! Distance metrics.
 
+use ndarray::ArrayView1;
 use num_traits::{Float, Zero};
 use std::ops::AddAssign;
 
 /// The type of a distance metric function.
-pub type Distance<A> = fn(&[A], &[A]) -> A;
+pub type Distance<A> = fn(&ArrayView1<A>, &ArrayView1<A>) -> A;
 
 /// Euclidean distance before taking a squre root. Used as a lightweight version
 /// of [`euclidean`] for relative comparisions.
 /// [`euclidean`]: #method.euclidean
-pub fn euclidean_reduced<A>(x1: &[A], x2: &[A]) -> A
+#[must_use]
+pub fn euclidean_reduced<A>(x1: &ArrayView1<A>, x2: &ArrayView1<A>) -> A
 where
     A: Float + Zero + AddAssign,
 {
@@ -23,7 +25,8 @@ where
 }
 
 /// Euclidean distance metric.
-pub fn euclidean<A>(x1: &[A], x2: &[A]) -> A
+#[must_use]
+pub fn euclidean<A>(x1: &ArrayView1<A>, x2: &ArrayView1<A>) -> A
 where
     A: Float + Zero + AddAssign,
 {