+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- `BallTree` and `VantagePointTree` accept an ndarray as a point.
+
 ## [0.4.0] - 2020-06-01
 
 ### Added
@@ -44,6 +50,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - A ball tree data structure to find nearest neighbors.
 
+[Unreleased]: https://github.com/petabi/petal-neighbors/compare/0.4.0...master
 [0.4.0]: https://github.com/petabi/petal-neighbors/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/petabi/petal-neighbors/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/petabi/petal-neighbors/compare/0.1.0...0.2.0